pyproject.toml
[project]
name = "text-editor"
version = "2.0.0"
description = "Rich text editor — FastAPI backend + HTML/JS frontend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "cloudinary>=1.40.0",
    "python-dotenv>=1.0.0",
]
